+++ src/main/java/me/pajic/cave_fog_stabilizer/platform/neoforge/NeoforgePlatform.java
@@ -1,0 +1,23 @@
+package me.pajic.cave_fog_stabilizer.platform.neoforge;
+
+//? if neoforge {
+
+/*import me.pajic.cave_fog_stabilizer.platform.Platform;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.FMLPaths;
+
+import java.nio.file.Path;
+
+public class NeoforgePlatform implements Platform {
+
+	@Override
+	public Path configDir() {
+		return FMLPaths.CONFIGDIR.get();
+	}
+
+	@Override
+	public boolean isModLoaded(String modId) {
+		return ModList.get().isLoaded(modId);
+	}
+}
+*///?}